root/workdir/vigil/requirements.txt
# VIGIL - Dependencies
# ====================

# Core AI
openai>=1.0.0
anthropic>=0.18.0

# Voice
SpeechRecognition>=3.10.0
pyttsx3>=2.90
elevenlabs>=1.0.0
pygame>=2.5.0
pyaudio>=0.2.14

# Utilities
requests>=2.31.0
python-dotenv>=1.0.0
pydantic>=2.0.0

# Optional: For Poe API access to Gemini
# fastapi-poe>=0.0.36

# Note: PyAudio may require additional setup on Windows
# If pip install pyaudio fails, try:
# pip install pipwin
# pipwin install pyaudio
